requirements.txt
line-bot-sdk
flask
pybooru
praw
deviantart